# openapi.yaml
openapi: 3.0.3
info:
  title: Meeting Service API
  description: API for scheduling meetings with participant availability tracking and slot recommendations
  version: 1.0.0
servers:
  - url: http://localhost:8080
    description: Local development server
paths:
...
  /preferred-slots:
    post:
      summary: Create preferred slot
      description: Create a new preferred time slot for a user (identified by email)
      operationId: createPreferredSlot
      tags:
        - Preferred Slots
      requestBody:
        required: true
        content:
          application/json:
            schema:
              $ref: '#/components/schemas/CreatePreferredSlotRequest'
      responses:
        '201':
          description: Preferred slot created successfully
          content:
            application/json:
              schema:
                $ref: '#/components/schemas/PreferredSlot'
        '400':
          description: Invalid request
          content:
            application/json:
              schema:
                $ref: '#/components/schemas/ErrorResponse'
        '500':
          description: Internal server error
          content:
            application/json:
              schema:
                $ref: '#/components/schemas/ErrorResponse'
  /preferred-slots/email/{email}:
    get:
      summary: Get preferred slots by email
      description: Get all preferred time slots for a user by their email address
      operationId: getPreferredSlotsByEmail
      tags:
        - Preferred Slots
      parameters:
        - name: email
          in: path
          required: true
          description: User email address
          schema:
            type: string
            format: email
      responses:
        '200':
          description: List of preferred slots
          content:
            application/json:
              schema:
                type: object
                properties:
                  preferred_slots:
                    type: array
                    items:
                      $ref: '#/components/schemas/PreferredSlot'
        '400':
          description: Invalid email
          content:
            application/json:
              schema:
                $ref: '#/components/schemas/ErrorResponse'
        '500':
          description: Internal server error
          content:
            application/json:
              schema:
                $ref: '#/components/schemas/ErrorResponse'
  /preferred-slots/{id}:
    put:
      summary: Update preferred slot
      description: Update an existing preferred time slot
      operationId: updatePreferredSlot
      tags:
        - Preferred Slots
      parameters:
        - $ref: '#/components/parameters/PreferredSlotId'
      requestBody:
        required: true
        content:
          application/json:
            schema:
              $ref: '#/components/schemas/UpdatePreferredSlotRequest'
      responses:
        '200':
          description: Preferred slot updated successfully
          content:
            application/json:
              schema:
                $ref: '#/components/schemas/PreferredSlot'
        '400':
          description: Invalid request
          content:
            application/json:
              schema:
                $ref: '#/components/schemas/ErrorResponse'
        '404':
          description: Preferred slot not found
          content:
            application/json:
              schema:
                $ref: '#/components/schemas/ErrorResponse'
        '500':
          description: Internal server error
          content:
            application/json:
              schema:
                $ref: '#/components/schemas/ErrorResponse'

    delete:
      summary: Delete preferred slot
      description: Delete a preferred time slot
      operationId: deletePreferredSlot
      tags:
        - Preferred Slots
      parameters:
        - $ref: '#/components/parameters/PreferredSlotId'
      responses:
        '204':
          description: Preferred slot deleted successfully
        '400':
          description: Invalid slot ID
          content:
            application/json:
              schema:
                $ref: '#/components/schemas/ErrorResponse'
        '404':
          description: Preferred slot not found
          content:
            application/json:
              schema:
                $ref: '#/components/schemas/ErrorResponse'
        '500':
          description: Internal server error
          content:
            application/json:
              schema:
                $ref: '#/components/schemas/ErrorResponse'
components:
  parameters:
...
    PreferredSlotId:
      name: id
      in: path
      required: true
      description: Preferred Slot UUID
      schema:
        type: string
        format: uuid
  schemas:
...
    ErrorResponse:
      type: object
      properties:
        error:
          type: string
...
    PreferredSlot:
      type: object
      properties:
        id:
          type: string
          format: uuid
        email:
          type: string
          format: email
        start_time:
          type: string
          format: date-time
        end_time:
          type: string
          format: date-time
        timezone:
          type: string
          description: IANA timezone identifier (e.g., "America/New_York")
        day_of_week:
          type: integer
          minimum: 0
          maximum: 6
          description: Day of week (0=Sunday, 6=Saturday). If set, slot applies only on this day.
        created_at:
          type: string
          format: date-time
        updated_at:
          type: string
          format: date-time
    CreatePreferredSlotRequest:
      type: object
      required:
        - email
        - start_time
        - end_time
        - timezone
      properties:
        email:
          type: string
          format: email
        start_time:
          type: string
          description: Start time in format "2006-01-02T15:04:05" or RFC3339
        end_time:
          type: string
          description: End time in format "2006-01-02T15:04:05" or RFC3339
        timezone:
          type: string
          description: IANA timezone identifier (e.g., "America/New_York")
        day_of_week:
          type: integer
          minimum: 0
          maximum: 6
          description: Day of week (0=Sunday, 6=Saturday). If set, slot applies only on this day.
    UpdatePreferredSlotRequest:
      type: object
      properties:
        start_time:
          type: string
          description: Start time in format "2006-01-02T15:04:05" or RFC3339
        end_time:
          type: string
          description: End time in format "2006-01-02T15:04:05" or RFC3339
        timezone:
          type: string
          description: IANA timezone identifier (e.g., "America/New_York")
        day_of_week:
          type: integer
          minimum: 0
          maximum: 6
          description: Day of week (0=Sunday, 6=Saturday). If set, slot applies only on this day.
...